Now using unsigned integers for buddhabrot heatmap, will be used for atomic addition

+++ shaders/2D/Fractals/buddhabrotRender.fs
@@ -6,7 +6,7 @@
 /*<bufferType>renderInput</bufferType>*/
 layout(std430, binding=1) buffer renderInput
 {
-    vec4 points[];
+    uvec4 points[];
 };
 </buffers>
